* Interface: Improved interface for user. * Code: Fixed logical error in File class, where the last command loaded into memory from the CIF file was cleaned, even when was not intended to be to. + Code: Added public static member function to the File class to validate a single string. + Code: Added public static member function to the File class to clean a single string. * Code: Fixed various warnings when compiling the library using flags to improve correcteness of code (-Wall -Werror -Wextra -pedantic -ansi). * Code: Header files renamed from "*.h" to "*.hh". This change is done to make more evident that they are headers of C++, not of C. * CMake: When installing the library into the system, only one header is copied into /usr/local/include/, named "opencif", instead of the folder hierarchy used previously.

+++ src/finitestatemachine/state.cc
@@ -20,7 +20,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */ 
 
-# include "state.h"
+# include "state.hh"
 
 /*
  * Default constructor. Initialice the vector of states.
@@ -55,9 +55,9 @@ void OpenCIF::State::reset ( void )
  */
 void OpenCIF::State::addOptions ( const std::string& new_options , const int& exit_state )
 {
-   for ( int i = 0; i < new_options.size (); i++ )
+   for ( unsigned int i = 0; i < new_options.size (); i++ )
    {
-      state_options[ new_options[ i ] ] = exit_state;
+      state_options[ (int)(new_options[ i ]) ] = exit_state;
    }
    
    return;
@@ -68,5 +68,5 @@ void OpenCIF::State::addOptions ( const std::string& new_options , const int& ex
  */
 int OpenCIF::State::operator[] ( const char& input_char )
 {
-   return ( state_options[ input_char ] );
+   return ( state_options[ (int)input_char ] );
 }
